Modified Repository and removed PlayerRepository Interface
