+++ src/main/java/com/hishab/app/repository/PlayerRepository.java
@@ -2,10 +2,31 @@ package com.hishab.app.repository;
 
 import java.util.List;
 
+import org.springframework.stereotype.Component;
+
+import com.hishab.app.exceptionhandling.PlayerOverflowException;
 import com.hishab.app.model.Player;
+import com.hishab.app.persistence.PersistenceDB;
 
-public interface PlayerRepository {
+@Component
+public class PlayerRepository {
 
-	Player addPlayer(Player player);
-	List<Player> getPlayers();
+	public Player addPlayer(Player player) {
+
+		Player newPlayer = new Player();
+
+		if (PersistenceDB.playerList.size() <= 3) {
+			newPlayer.setPlayerId(player.getPlayerId());
+			newPlayer.setName(player.getName());
+			newPlayer.setAge(player.getAge());
+			PersistenceDB.playerList.add(newPlayer);
+		} else {
+			throw new PlayerOverflowException("Cross the Player quota. Please wait for next Game");
+		}
+		return newPlayer;
+	}
+
+	public List<Player> getPlayers() {
+		return PersistenceDB.playerList;
+	}
 }
