+++ src/test/java/com/hishab/app/HishabAppApplicationTests.java
@@ -2,12 +2,13 @@ package com.hishab.app;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
+@ActiveProfiles("test")
 @SpringBootTest
 class HishabAppApplicationTests {
 
 	@Test
 	void contextLoads() {
 	}
-
 }
