+++ src/test/java/com/hishab/app/client/DiceFeignClientTest.java
@@ -2,9 +2,6 @@ package com.hishab.app.client;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
